enh: rename ``eddymotion._sklearn`` -> ``eddymotion.gpr``
* rename the module so that it is picked up by autodoc.
* also fixes minor errors and improves docs.
